Combine Calendar and Celebrations into one settings page
The Calendar and Celebrations pages overlapped heavily — Calendar's
Birthdays / Public-holidays feature toggles drove the exact same prefs
as Celebrations' "Calendar Holidays / Calendar Birthdays" source rows,
and each page kept its own duplicate READ_CALENDAR plumbing. Merge them
into a single "Calendar" page (subhead "Events, Holidays, and
Birthdays") and clean up the structure so it reads top-down:

  1. Calendar permission — the READ_CALENDAR master toggle
  2. From your calendar — Evening tie-ins / Birthdays / Public holidays
  3. Built-in holidays — Region / Location / Global / Funny + a
     "Browse by country" expand row that reveals the per-country
     drill-down catalogue (collapsed by default so the page top stays
     scannable)
  4. Calendar Holidays / Birthdays listings — the next year of
     detected events from the user's synced calendars

One shared permission state now backs both the feature toggles and the
listings. The duplicated "Calendar Holidays/Birthdays" source rows are
dropped (the feature toggles already own those prefs). The built-in
holiday catalogue stays independent of calendar permission — country
toggles and the drill-down work without calendar access.

Renames: "What to use it for" → "From your calendar" (makes clear those
are the calendar-gated features, contrasted with the next card's
"Built-in holidays" sources). "Default holidays" was on the way to the
same place via the earlier PR's #625 review thread — settled on
"Built-in holidays" to contrast with calendar-detected.

The Today screen's "celebration themes" card and the old Celebrations
menu entry both now route to this page; the separate Celebrations menu
row, HolidaysPage, and HolidaysDest route are removed.

No change to what leaves the device — purely a UI consolidation;
calendar reads, permission handling, and the data sent to weather / TTS
are all unchanged.

Snapshot: settings_calendar regenerated to the merged + collapsed-
catalogue layout; settings_holidays removed.
