+++ app/src/main/kotlin/app/clothescast/ui/settings/HolidaySettings.kt
@@ -25,6 +25,7 @@ import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
@@ -33,6 +34,7 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
@@ -43,11 +45,13 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.contentDescription
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.unit.dp
+import androidx.core.app.ActivityCompat
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import app.clothescast.R
 import app.clothescast.calendar.CalendarPermission
+import app.clothescast.diag.findActivity
 import app.clothescast.core.domain.model.EventKind
 import app.clothescast.core.domain.model.HolidayCatalog
 import app.clothescast.core.domain.model.HolidayCountrySelection
@@ -56,6 +60,7 @@ import app.clothescast.core.domain.model.HolidayOverride
 import app.clothescast.core.domain.model.HolidayTheme
 import app.clothescast.core.domain.model.UpcomingCalendarEvent
 import app.clothescast.ui.EdgeFadeOverlay
+import kotlinx.coroutines.launch
 import java.text.Collator
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
@@ -63,39 +68,47 @@ import java.time.format.FormatStyle
 import java.util.Locale
 
 /**
- * A top-of-page card with the country-source toggles (Home / Current
- * location / Global / Funny / All), followed by a Global collapsible
- * listing the four universal holidays (Christmas, NYE, Halloween,
- * Valentine's), a Funny collapsible listing the playful observances
- * (Talk Like a Pirate Day), then one collapsible per ISO country
- * listing that country's holidays, and finally an "All" collapsible
- * showing every holiday in the catalogue flat for power-user search.
+ * The Calendar settings page. Top to bottom: a "Use my calendar" permission
+ * master toggle, a "What to use it for" card with the three calendar-sourced
+ * feature toggles (Evening tie-ins, Birthdays, Public holidays), the
+ * country-source toggles (Region / Location / Global / Funny / All), the two
+ * upcoming-celebration listings (Calendar Holidays / Birthdays), and finally
+ * the curated holiday catalogue: a Global collapsible listing the four
+ * universal holidays (Christmas, NYE, Halloween, Valentine's), a Funny
+ * collapsible listing the playful observances (Talk Like a Pirate Day), one
+ * collapsible per ISO country, and an "All" collapsible showing every holiday
+ * flat for power-user search.
  *
- * Both section headers and individual holiday rows carry a tri-state
- * dropdown (`Auto` / `On` / `Off`). The `Auto` label includes the
- * current resolution ("Auto (on)" / "Auto (off)") so the user can see
- * what the country picker (or the country override) is doing without
- * flipping it in their head.
+ * The calendar permission toggle gates only the calendar-sourced features and
+ * the upcoming-event listings; the curated catalogue (country sources +
+ * collapsibles) is independent of it and stays usable without calendar access.
  *
- * Section headers also show an `(n/m)` summary — holidays currently
- * active / total for that bucket.
+ * Catalogue section headers and individual holiday rows carry a tri-state
+ * dropdown (`Auto` / `On` / `Off`). The `Auto` label includes the current
+ * resolution ("Auto (on)" / "Auto (off)") so the user can see what the country
+ * picker (or the country override) is doing without flipping it in their head.
+ * Section headers also show an `(n/m)` summary — holidays currently active /
+ * total for that bucket.
  *
  * Resource IDs are looked up via [LocalContext]'s `getIdentifier` so the
  * theme catalogue can live in `:core:domain` without depending on `R`.
  * Missing translations fall back to the raw enum name / ISO code.
  */
 @Composable
-internal fun HolidaysContent(
+internal fun CalendarContent(
     holidayCountrySelection: HolidayCountrySelection,
     holidayOverrides: Map<HolidayId, HolidayOverride>,
     effectiveEnabledHolidayCountries: Set<String>,
     localeCountry: String?,
     weatherLocationCountry: String?,
     calendarEnabled: Boolean,
+    useCalendarEvents: Boolean,
     themeFromCalendarHolidays: Boolean,
     themeFromCalendarBirthdays: Boolean,
     calendarCelebrations: List<UpcomingCalendarEvent>?,
     padding: PaddingValues,
+    onSetCalendarEnabled: (Boolean) -> Unit,
+    onSetUseCalendarEvents: (Boolean) -> Unit,
     onSetCountryHome: (Boolean) -> Unit,
     onSetCountryCurrent: (Boolean) -> Unit,
     onSetCountryGlobal: (Boolean) -> Unit,
@@ -109,36 +122,92 @@ internal fun HolidaysContent(
     onLoadCalendarCelebrations: () -> Unit,
     onNavigateToRegionSettings: () -> Unit,
     onNavigateToLocationSettings: () -> Unit,
-    onNavigateToCalendarSettings: () -> Unit,
 ) {
     val scrollState = rememberScrollState()
+    val coroutineScope = rememberCoroutineScope()
     val context = LocalContext.current
     val uiLocale = remember(context.resources.configuration) {
         context.resources.configuration.locales.get(0) ?: Locale.getDefault()
     }
 
-    // Calendar-permission state for the upcoming-celebrations listings below.
-    // Tracked once here (shared by both the Holidays and Birthdays sections)
-    // and re-checked on resume so a grant/revoke from system Settings flips the
-    // sections between their prompt and the live list. The listing reads only
-    // when the master calendar switch is on AND permission is granted — turning
-    // the whole calendar feature off (in Calendar settings) stops the read.
-    var calendarPermissionGranted by remember { mutableStateOf(CalendarPermission.isGranted(context)) }
+    // One source of permission truth for the whole page — the master toggle, the
+    // three feature toggles, and the two upcoming-celebration listings all read
+    // it. Re-checked on resume so a grant/revoke from system Settings is
+    // reflected without leaving the page; a revoke flips the master switch off
+    // (and pings consumers to refresh) so we don't keep claiming calendar access
+    // we no longer have. The curated holiday catalogue below is deliberately NOT
+    // gated on this — built-in holidays work without calendar permission.
+    var permissionGranted by remember { mutableStateOf(CalendarPermission.isGranted(context)) }
+    // True once a permission request comes back denied *and* the system won't
+    // show the rationale dialog again ("don't ask again" / permanently denied).
+    // In that state tapping a toggle does nothing — the OS suppresses the prompt
+    // — so the only way back is the app's system-settings screen, which is the
+    // sole reason we surface a button at all.
+    var permanentlyDenied by remember { mutableStateOf(false) }
+    val currentCalendarEnabled by rememberUpdatedState(calendarEnabled)
+    val currentOnSetCalendarEnabled by rememberUpdatedState(onSetCalendarEnabled)
+    val currentOnRechecked by rememberUpdatedState(onCalendarPermissionRechecked)
     val lifecycleOwner = LocalLifecycleOwner.current
     DisposableEffect(lifecycleOwner) {
         val observer = LifecycleEventObserver { _, event ->
             if (event == Lifecycle.Event.ON_RESUME) {
-                calendarPermissionGranted = CalendarPermission.isGranted(context)
+                val granted = CalendarPermission.isGranted(context)
+                permissionGranted = granted
+                if (granted) {
+                    permanentlyDenied = false
+                } else if (currentCalendarEnabled) {
+                    currentOnSetCalendarEnabled(false)
+                    currentOnRechecked()
+                }
             }
         }
         lifecycleOwner.lifecycle.addObserver(observer)
         onDispose { lifecycleOwner.lifecycle.removeObserver(observer) }
     }
-    val calendarPermissionLauncher = rememberLauncherForActivityResult(
+
+    // Single permission launcher with a pending action: whichever toggle asked
+    // to turn on runs its enable callback once permission is granted. The VM
+    // flips the master switch on when any sub-feature is enabled, so enabling a
+    // sub-feature from scratch (master off) prompts once and lights up both.
+    var pendingEnable by remember { mutableStateOf<(() -> Unit)?>(null) }
+    val launcher = rememberLauncherForActivityResult(
         contract = ActivityResultContracts.RequestPermission(),
-    ) { granted -> calendarPermissionGranted = granted }
-    LaunchedEffect(calendarEnabled, calendarPermissionGranted) {
-        if (calendarEnabled && calendarPermissionGranted) onLoadCalendarCelebrations()
+    ) { granted ->
+        permissionGranted = granted
+        if (granted) {
+            permanentlyDenied = false
+            pendingEnable?.invoke()
+            currentOnRechecked()
+        } else {
+            // Denied. If the system would still show the rationale dialog the
+            // user can just try the toggle again; if not, they're permanently
+            // denied and we point them at system settings.
+            val activity = context.findActivity()
+            permanentlyDenied = activity != null &&
+                !ActivityCompat.shouldShowRequestPermissionRationale(
+                    activity,
+                    CalendarPermission.MANIFEST_PERMISSION,
+                )
+        }
+        pendingEnable = null
+    }
+    val requestThenEnable: (() -> Unit) -> Unit = { action ->
+        if (permissionGranted) action() else {
+            pendingEnable = action
+            launcher.launch(CalendarPermission.MANIFEST_PERMISSION)
+        }
+    }
+    // The listings' in-place "grant permission" prompt requests access without a
+    // pending feature enable — granting just lets the read run.
+    val requestPermissionForListing: () -> Unit = {
+        if (!permissionGranted) launcher.launch(CalendarPermission.MANIFEST_PERMISSION)
+    }
+    // Master off → the listing prompt scrolls back up to the toggle (same page)
+    // rather than deep-linking elsewhere.
+    val scrollToTop: () -> Unit = { coroutineScope.launch { scrollState.animateScrollTo(0) } }
+
+    LaunchedEffect(calendarEnabled, permissionGranted) {
+        if (calendarEnabled && permissionGranted) onLoadCalendarCelebrations()
     }
     val calendarHolidays = calendarCelebrations?.filter { it.kind == EventKind.PUBLIC_HOLIDAY }
     val calendarBirthdays = calendarCelebrations?.filter { it.kind == EventKind.BIRTHDAY }
@@ -188,6 +257,12 @@ internal fun HolidaysContent(
         sortedCatalog.map { it.second }
     }
 
+    // Tucks the per-country / Global / Funny / All collapsibles behind a
+    // single expand row at the bottom of the Built-in holidays card.
+    // Collapsed by default so the page top stays scannable; expansion state
+    // is saved across config changes for the power-user case.
+    var browseCountriesExpanded by rememberSaveable { mutableStateOf(false) }
+
     EdgeFadeOverlay(
         scrollState = scrollState,
         modifier = Modifier.padding(padding),
@@ -200,213 +275,268 @@ internal fun HolidaysContent(
                 .padding(horizontal = 16.dp, vertical = 8.dp),
             verticalArrangement = Arrangement.spacedBy(12.dp),
         ) {
-            // Sources card — five holiday-source toggles, each in its own
-            // row. Region / Location / Calendar Holidays / Calendar Birthdays
-            // carry a deep-link to the corresponding settings page. Global
-            // is the universal-holiday bucket (Christmas, NYE, Halloween,
-            // Valentine's) with no dedicated settings target.
-            Card(modifier = Modifier.fillMaxWidth()) {
-                Column(
-                    modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp),
-                    verticalArrangement = Arrangement.spacedBy(4.dp),
+            SectionCard(title = stringResource(R.string.settings_calendar_features_title)) {
+                // Master toggle for calendar access. Gating: when this is off,
+                // the three feature rows below render as off (their `checked`
+                // gates on `calendarEnabled && permissionGranted`) and a
+                // request-permission tap on any of them re-enables the master
+                // before flipping the feature pref. On: the toggle itself is
+                // the permission prompt — no separate "grant" button. Off: we
+                // just stop reading in-app (the *Active gates); we deliberately
+                // don't relinquish the OS permission — there's no immediate
+                // API for it, and `revokeSelfPermissionsOnKill` is API 33+
+                // (minSdk is 31) and only takes effect after the process is
+                // killed, so it'd add a version-gated, surprising code path
+                // for no real benefit until we bump minSdk to Android 13+.
+                CalendarFeatureRow(
+                    label = stringResource(R.string.settings_calendar_master),
+                    description = stringResource(R.string.settings_calendar_master_description),
+                    checked = calendarEnabled && permissionGranted,
+                    onToggle = { wantsOn ->
+                        if (wantsOn) requestThenEnable { onSetCalendarEnabled(true) }
+                        else onSetCalendarEnabled(false)
+                    },
+                )
+                if (permanentlyDenied && !permissionGranted) {
+                    Text(
+                        text = stringResource(R.string.settings_calendar_open_settings),
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.error,
+                    )
+                    TextButton(
+                        onClick = { openAppDetails(context) },
+                        modifier = Modifier.fillMaxWidth(),
+                    ) { Text(stringResource(R.string.settings_calendar_open_system_settings)) }
+                }
+                CalendarFeatureRow(
+                    label = stringResource(R.string.settings_calendar_evening_tie_ins),
+                    description = stringResource(R.string.settings_calendar_evening_tie_ins_description),
+                    checked = calendarEnabled && useCalendarEvents && permissionGranted,
+                    onToggle = { wantsOn ->
+                        if (wantsOn) requestThenEnable { onSetUseCalendarEvents(true) }
+                        else onSetUseCalendarEvents(false)
+                    },
+                )
+                CalendarFeatureRow(
+                    label = stringResource(R.string.settings_calendar_birthdays),
+                    description = stringResource(R.string.settings_calendar_birthdays_description),
+                    checked = calendarEnabled && themeFromCalendarBirthdays && permissionGranted,
+                    onToggle = { wantsOn ->
+                        if (wantsOn) requestThenEnable { onSetThemeFromCalendarBirthdays(true) }
+                        else onSetThemeFromCalendarBirthdays(false)
+                    },
+                )
+                CalendarFeatureRow(
+                    label = stringResource(R.string.settings_calendar_public_holidays),
+                    description = stringResource(R.string.settings_calendar_public_holidays_description),
+                    checked = calendarEnabled && themeFromCalendarHolidays && permissionGranted,
+                    onToggle = { wantsOn ->
+                        if (wantsOn) requestThenEnable { onSetThemeFromCalendarHolidays(true) }
+                        else onSetThemeFromCalendarHolidays(false)
+                    },
+                )
+            }
+
+            // Curated celebration sources — country buckets that drive the
+            // built-in holiday catalogue below. Independent of calendar
+            // permission. Region / Location carry a deep-link to the
+            // corresponding settings page; Global / Funny have no dedicated
+            // target.
+            SectionCard(title = stringResource(R.string.settings_holidays_sources_title)) {
+                SourceRow(
+                    label = stringResource(
+                        R.string.settings_holidays_source_region,
+                        localeCountry?.let { resolveCountryDisplayName(context, uiLocale, it) }
+                            ?: stringResource(R.string.settings_holiday_country_unknown),
+                    ),
+                    checked = holidayCountrySelection.home,
+                    onCheckedChange = onSetCountryHome,
+                    linkLabel = stringResource(R.string.settings_holidays_link_settings),
+                    linkContentDescription = stringResource(R.string.settings_holidays_link_region_settings_a11y),
+                    onLinkClick = onNavigateToRegionSettings,
+                )
+                SourceRow(
+                    label = stringResource(
+                        R.string.settings_holidays_source_location,
+                        weatherLocationCountry?.let { resolveCountryDisplayName(context, uiLocale, it) }
+                            ?: stringResource(R.string.settings_holiday_country_unknown),
+                    ),
+                    checked = holidayCountrySelection.current,
+                    onCheckedChange = onSetCountryCurrent,
+                    linkLabel = stringResource(R.string.settings_holidays_link_settings),
+                    linkContentDescription = stringResource(R.string.settings_holidays_link_location_settings_a11y),
+                    onLinkClick = onNavigateToLocationSettings,
+                )
+                SourceRow(
+                    label = stringResource(R.string.settings_holidays_source_global),
+                    checked = holidayCountrySelection.global,
+                    onCheckedChange = onSetCountryGlobal,
+                )
+                SourceRow(
+                    label = stringResource(R.string.settings_holidays_source_funny),
+                    checked = holidayCountrySelection.funny,
+                    onCheckedChange = onSetCountryFunny,
+                )
+                // Legacy "All countries" toggle, conditionally rendered so
+                // users who'd enabled the previous UI's All checkbox can
+                // turn it off from here. With `all=true` the resolver
+                // short-circuits every country on, making Region/Location/
+                // Global toggles below visually-meaningless — surface a
+                // clear opt-out instead of forcing app-data wipe.
+                if (holidayCountrySelection.all) {
+                    SourceRow(
+                        label = stringResource(R.string.settings_holidays_source_all),
+                        checked = true,
+                        onCheckedChange = onSetCountryAll,
+                    )
+                }
+                // Expand row that reveals the per-country / Global / Funny /
+                // All catalogue below the card. Off-page by default so the
+                // top of Calendar stays scannable.
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    verticalAlignment = Alignment.CenterVertically,
                 ) {
-                    SourceRow(
-                        label = stringResource(
-                            R.string.settings_holidays_source_region,
-                            localeCountry?.let { resolveCountryDisplayName(context, uiLocale, it) }
-                                ?: stringResource(R.string.settings_holiday_country_unknown),
-                        ),
-                        checked = holidayCountrySelection.home,
-                        onCheckedChange = onSetCountryHome,
-                        linkLabel = stringResource(R.string.settings_holidays_link_settings),
-                        linkContentDescription = stringResource(R.string.settings_holidays_link_region_settings_a11y),
-                        onLinkClick = onNavigateToRegionSettings,
+                    Text(
+                        text = stringResource(R.string.settings_holidays_browse_by_country),
+                        style = MaterialTheme.typography.bodyLarge,
+                        modifier = Modifier.weight(1f),
                     )
-                    SourceRow(
-                        label = stringResource(
-                            R.string.settings_holidays_source_location,
-                            weatherLocationCountry?.let { resolveCountryDisplayName(context, uiLocale, it) }
-                                ?: stringResource(R.string.settings_holiday_country_unknown),
-                        ),
-                        checked = holidayCountrySelection.current,
-                        onCheckedChange = onSetCountryCurrent,
-                        linkLabel = stringResource(R.string.settings_holidays_link_settings),
-                        linkContentDescription = stringResource(R.string.settings_holidays_link_location_settings_a11y),
-                        onLinkClick = onNavigateToLocationSettings,
-                    )
-                    SourceRow(
-                        label = stringResource(R.string.settings_holidays_source_global),
-                        checked = holidayCountrySelection.global,
-                        onCheckedChange = onSetCountryGlobal,
-                    )
-                    SourceRow(
-                        label = stringResource(R.string.settings_holidays_source_funny),
-                        checked = holidayCountrySelection.funny,
-                        onCheckedChange = onSetCountryFunny,
-                    )
-                    CalendarSourceRow(
-                        label = stringResource(R.string.settings_holidays_source_calendar_holidays),
-                        checked = calendarEnabled && themeFromCalendarHolidays,
-                        onSetChecked = onSetThemeFromCalendarHolidays,
-                        onPermissionRechecked = onCalendarPermissionRechecked,
-                        linkLabel = stringResource(R.string.settings_holidays_link_settings),
-                        linkContentDescription = stringResource(R.string.settings_holidays_link_calendar_settings_a11y),
-                        onLinkClick = onNavigateToCalendarSettings,
-                    )
-                    CalendarSourceRow(
-                        label = stringResource(R.string.settings_holidays_source_calendar_birthdays),
-                        checked = calendarEnabled && themeFromCalendarBirthdays,
-                        onSetChecked = onSetThemeFromCalendarBirthdays,
-                        onPermissionRechecked = onCalendarPermissionRechecked,
-                        linkLabel = stringResource(R.string.settings_holidays_link_settings),
-                        linkContentDescription = stringResource(R.string.settings_holidays_link_calendar_settings_a11y),
-                        onLinkClick = onNavigateToCalendarSettings,
-                    )
-                    // Legacy "All countries" toggle, conditionally rendered so
-                    // users who'd enabled the previous UI's All checkbox can
-                    // turn it off from here. With `all=true` the resolver
-                    // short-circuits every country on, making Region/Location/
-                    // Global toggles below visually-meaningless — surface a
-                    // clear opt-out instead of forcing app-data wipe.
-                    if (holidayCountrySelection.all) {
-                        SourceRow(
-                            label = stringResource(R.string.settings_holidays_source_all),
-                            checked = true,
-                            onCheckedChange = onSetCountryAll,
+                    IconButton(onClick = { browseCountriesExpanded = !browseCountriesExpanded }) {
+                        Icon(
+                            imageVector = if (browseCountriesExpanded) Icons.Filled.KeyboardArrowUp else Icons.Filled.KeyboardArrowDown,
+                            contentDescription = stringResource(
+                                if (browseCountriesExpanded) R.string.settings_holidays_collapse
+                                else R.string.settings_holidays_expand
+                            ),
                         )
                     }
                 }
             }
 
-            // Calendar Holidays / Birthdays listings — the next year of detected
-            // events from the user's synced calendars, collapsed by default and
-            // gated on READ_CALENDAR. When permission is missing each section
-            // offers an in-place grant prompt instead of a list.
-            CalendarCelebrationsSection(
-                title = stringResource(R.string.settings_holidays_source_calendar_holidays),
-                rememberKey = "holidays-calendar-holidays-section",
-                calendarEnabled = calendarEnabled,
-                permissionGranted = calendarPermissionGranted,
-                events = calendarHolidays,
-                emptyMessage = stringResource(R.string.settings_holidays_calendar_no_holidays),
-                uiLocale = uiLocale,
-                onRequestPermission = {
-                    calendarPermissionLauncher.launch(CalendarPermission.MANIFEST_PERMISSION)
-                },
-                onEnableCalendar = onNavigateToCalendarSettings,
-            )
-            CalendarCelebrationsSection(
-                title = stringResource(R.string.settings_holidays_source_calendar_birthdays),
-                rememberKey = "holidays-calendar-birthdays-section",
-                calendarEnabled = calendarEnabled,
-                permissionGranted = calendarPermissionGranted,
-                events = calendarBirthdays,
-                emptyMessage = stringResource(R.string.settings_holidays_calendar_no_birthdays),
-                uiLocale = uiLocale,
-                onRequestPermission = {
-                    calendarPermissionLauncher.launch(CalendarPermission.MANIFEST_PERMISSION)
-                },
-                onEnableCalendar = onNavigateToCalendarSettings,
-            )
-
-            val globalActiveCount = globalThemes.count { theme ->
-                theme.isActive(holidayOverrides, effectiveEnabledHolidayCountries)
-            }
-            val globalOverride = holidayCountrySelection.countryOverrides[HolidayCatalog.GLOBAL_COUNTRY]
-                ?: HolidayOverride.AUTO
-            val globalAutoOn = holidayCountrySelection.countryAutoEffective(
-                HolidayCatalog.GLOBAL_COUNTRY,
-                localeCountry,
-                weatherLocationCountry,
-            )
-            CollapsibleSection(
-                title = stringResource(R.string.settings_holiday_country_global_label),
-                summary = "$globalActiveCount/${globalThemes.size}",
-                rememberKey = "holidays-country-${HolidayCatalog.GLOBAL_COUNTRY}",
-                trailing = {
-                    OverrideDropdown(
-                        current = globalOverride,
-                        autoOn = globalAutoOn,
-                        onChange = { newState ->
-                            onSetCountryOverride(HolidayCatalog.GLOBAL_COUNTRY, newState)
-                        },
-                    )
-                },
-            ) {
-                globalThemes.forEach { theme ->
-                    HolidayOverrideRow(
-                        theme = theme,
-                        override = holidayOverrides[theme.id] ?: HolidayOverride.AUTO,
-                        autoOn = theme.countries.any { it in effectiveEnabledHolidayCountries },
-                        onChange = { newState -> onSetHolidayOverride(theme.id, newState) },
-                    )
+            // Per-country / Global / Funny / All catalogue — drill-down for
+            // the four source-row toggles above. Off-page by default and
+            // revealed via the "Browse by country" expand row on the
+            // Built-in holidays card. Each section's expansion state is
+            // independently saved via [CollapsibleSection]'s rememberKey, so
+            // toggling Browse-by-country off and on returns the user to the
+            // same sub-section they were drilling into.
+            if (browseCountriesExpanded) {
+                val globalActiveCount = globalThemes.count { theme ->
+                    theme.isActive(holidayOverrides, effectiveEnabledHolidayCountries)
                 }
-            }
-
-            val funnyActiveCount = funnyThemes.count { theme ->
-                theme.isActive(holidayOverrides, effectiveEnabledHolidayCountries)
-            }
-            val funnyOverride = holidayCountrySelection.countryOverrides[HolidayCatalog.FUNNY]
-                ?: HolidayOverride.AUTO
-            val funnyAutoOn = holidayCountrySelection.countryAutoEffective(
-                HolidayCatalog.FUNNY,
-                localeCountry,
-                weatherLocationCountry,
-            )
-            CollapsibleSection(
-                title = stringResource(R.string.settings_holiday_country_funny_label),
-                summary = "$funnyActiveCount/${funnyThemes.size}",
-                rememberKey = "holidays-country-${HolidayCatalog.FUNNY}",
-                trailing = {
-                    OverrideDropdown(
-                        current = funnyOverride,
-                        autoOn = funnyAutoOn,
-                        onChange = { newState ->
-                            onSetCountryOverride(HolidayCatalog.FUNNY, newState)
-                        },
-                    )
-                },
-            ) {
-                funnyThemes.forEach { theme ->
-                    HolidayOverrideRow(
-                        theme = theme,
-                        override = holidayOverrides[theme.id] ?: HolidayOverride.AUTO,
-                        autoOn = theme.countries.any { it in effectiveEnabledHolidayCountries },
-                        onChange = { newState -> onSetHolidayOverride(theme.id, newState) },
-                    )
-                }
-            }
-
-            // Pull the Region and Location countries (if any) to the top of
-            // the per-country list — the user's "own" countries first, then
-            // the rest of the catalog alphabetically.
-            val pinnedCountries = listOfNotNull(localeCountry?.uppercase(), weatherLocationCountry?.uppercase())
-                .filter { it in isoCountries }
-                .distinct()
-            val orderedCountries = pinnedCountries + isoCountries.filterNot { it in pinnedCountries }
-
-            orderedCountries.forEach { code ->
-                val themes = themesByCountry[code].orEmpty()
-                val activeCount = themes.count { theme -> theme.isActive(holidayOverrides, effectiveEnabledHolidayCountries) }
-                val countryOverride = holidayCountrySelection.countryOverrides[code] ?: HolidayOverride.AUTO
-                val countryAutoOn = holidayCountrySelection.countryAutoEffective(
-                    code,
+                val globalOverride = holidayCountrySelection.countryOverrides[HolidayCatalog.GLOBAL_COUNTRY]
+                    ?: HolidayOverride.AUTO
+                val globalAutoOn = holidayCountrySelection.countryAutoEffective(
+                    HolidayCatalog.GLOBAL_COUNTRY,
                     localeCountry,
                     weatherLocationCountry,
                 )
                 CollapsibleSection(
-                    title = resolveCountryDisplayName(context, uiLocale, code),
-                    summary = "$activeCount/${themes.size}",
-                    rememberKey = "holidays-country-$code",
+                    title = stringResource(R.string.settings_holiday_country_global_label),
+                    summary = "$globalActiveCount/${globalThemes.size}",
+                    rememberKey = "holidays-country-${HolidayCatalog.GLOBAL_COUNTRY}",
                     trailing = {
                         OverrideDropdown(
-                            current = countryOverride,
-                            autoOn = countryAutoOn,
-                            onChange = { newState -> onSetCountryOverride(code, newState) },
+                            current = globalOverride,
+                            autoOn = globalAutoOn,
+                            onChange = { newState ->
+                                onSetCountryOverride(HolidayCatalog.GLOBAL_COUNTRY, newState)
+                            },
                         )
                     },
                 ) {
-                    themes.forEach { theme ->
+                    globalThemes.forEach { theme ->
+                        HolidayOverrideRow(
+                            theme = theme,
+                            override = holidayOverrides[theme.id] ?: HolidayOverride.AUTO,
+                            autoOn = theme.countries.any { it in effectiveEnabledHolidayCountries },
+                            onChange = { newState -> onSetHolidayOverride(theme.id, newState) },
+                        )
+                    }
+                }
+
+                val funnyActiveCount = funnyThemes.count { theme ->
+                    theme.isActive(holidayOverrides, effectiveEnabledHolidayCountries)
+                }
+                val funnyOverride = holidayCountrySelection.countryOverrides[HolidayCatalog.FUNNY]
+                    ?: HolidayOverride.AUTO
+                val funnyAutoOn = holidayCountrySelection.countryAutoEffective(
+                    HolidayCatalog.FUNNY,
+                    localeCountry,
+                    weatherLocationCountry,
+                )
+                CollapsibleSection(
+                    title = stringResource(R.string.settings_holiday_country_funny_label),
+                    summary = "$funnyActiveCount/${funnyThemes.size}",
+                    rememberKey = "holidays-country-${HolidayCatalog.FUNNY}",
+                    trailing = {
+                        OverrideDropdown(
+                            current = funnyOverride,
+                            autoOn = funnyAutoOn,
+                            onChange = { newState ->
+                                onSetCountryOverride(HolidayCatalog.FUNNY, newState)
+                            },
+                        )
+                    },
+                ) {
+                    funnyThemes.forEach { theme ->
+                        HolidayOverrideRow(
+                            theme = theme,
+                            override = holidayOverrides[theme.id] ?: HolidayOverride.AUTO,
+                            autoOn = theme.countries.any { it in effectiveEnabledHolidayCountries },
+                            onChange = { newState -> onSetHolidayOverride(theme.id, newState) },
+                        )
+                    }
+                }
+
+                // Pull the Region and Location countries (if any) to the top
+                // of the per-country list — the user's "own" countries first,
+                // then the rest of the catalog alphabetically.
+                val pinnedCountries = listOfNotNull(localeCountry?.uppercase(), weatherLocationCountry?.uppercase())
+                    .filter { it in isoCountries }
+                    .distinct()
+                val orderedCountries = pinnedCountries + isoCountries.filterNot { it in pinnedCountries }
+
+                orderedCountries.forEach { code ->
+                    val themes = themesByCountry[code].orEmpty()
+                    val activeCount = themes.count { theme -> theme.isActive(holidayOverrides, effectiveEnabledHolidayCountries) }
+                    val countryOverride = holidayCountrySelection.countryOverrides[code] ?: HolidayOverride.AUTO
+                    val countryAutoOn = holidayCountrySelection.countryAutoEffective(
+                        code,
+                        localeCountry,
+                        weatherLocationCountry,
+                    )
+                    CollapsibleSection(
+                        title = resolveCountryDisplayName(context, uiLocale, code),
+                        summary = "$activeCount/${themes.size}",
+                        rememberKey = "holidays-country-$code",
+                        trailing = {
+                            OverrideDropdown(
+                                current = countryOverride,
+                                autoOn = countryAutoOn,
+                                onChange = { newState -> onSetCountryOverride(code, newState) },
+                            )
+                        },
+                    ) {
+                        themes.forEach { theme ->
+                            HolidayOverrideRow(
+                                theme = theme,
+                                override = holidayOverrides[theme.id] ?: HolidayOverride.AUTO,
+                                autoOn = theme.countries.any { it in effectiveEnabledHolidayCountries },
+                                onChange = { newState -> onSetHolidayOverride(theme.id, newState) },
+                            )
+                        }
+                    }
+                }
+
+                val allActiveCount = allThemes.count { theme -> theme.isActive(holidayOverrides, effectiveEnabledHolidayCountries) }
+                CollapsibleSection(
+                    title = stringResource(R.string.settings_holiday_all_section_title),
+                    summary = "$allActiveCount/${allThemes.size}",
+                    rememberKey = "holidays-all-section",
+                ) {
+                    allThemes.forEach { theme ->
                         HolidayOverrideRow(
                             theme = theme,
                             override = holidayOverrides[theme.id] ?: HolidayOverride.AUTO,
@@ -418,28 +548,39 @@ internal fun HolidaysContent(
             }
 
             // TODO(celebrations-v2): give each row in the Calendar Holidays /
-            // Birthdays listings above a per-event override dropdown (same
+            // Birthdays listings below a per-event override dropdown (same
             // pattern as the country sections) so the user can mute a specific
             // event ("Boxing Day" they don't celebrate; a noisy birthday import)
             // without disabling the whole source. Needs a CalendarEvent → stable
             // key scheme that survives event-recurrence renames, persisted
             // alongside holidayOverrides.
 
-            val allActiveCount = allThemes.count { theme -> theme.isActive(holidayOverrides, effectiveEnabledHolidayCountries) }
-            CollapsibleSection(
-                title = stringResource(R.string.settings_holiday_all_section_title),
-                summary = "$allActiveCount/${allThemes.size}",
-                rememberKey = "holidays-all-section",
-            ) {
-                allThemes.forEach { theme ->
-                    HolidayOverrideRow(
-                        theme = theme,
-                        override = holidayOverrides[theme.id] ?: HolidayOverride.AUTO,
-                        autoOn = theme.countries.any { it in effectiveEnabledHolidayCountries },
-                        onChange = { newState -> onSetHolidayOverride(theme.id, newState) },
-                    )
-                }
-            }
+            // Calendar Holidays / Birthdays listings — the next year of detected
+            // events from the user's synced calendars, collapsed by default and
+            // gated on READ_CALENDAR. When permission is missing each section
+            // offers an in-place grant prompt instead of a list.
+            CalendarCelebrationsSection(
+                title = stringResource(R.string.settings_holidays_source_calendar_holidays),
+                rememberKey = "holidays-calendar-holidays-section",
+                calendarEnabled = calendarEnabled,
+                permissionGranted = permissionGranted,
+                events = calendarHolidays,
+                emptyMessage = stringResource(R.string.settings_holidays_calendar_no_holidays),
+                uiLocale = uiLocale,
+                onRequestPermission = requestPermissionForListing,
+                onEnableCalendar = scrollToTop,
+            )
+            CalendarCelebrationsSection(
+                title = stringResource(R.string.settings_holidays_source_calendar_birthdays),
+                rememberKey = "holidays-calendar-birthdays-section",
+                calendarEnabled = calendarEnabled,
+                permissionGranted = permissionGranted,
+                events = calendarBirthdays,
+                emptyMessage = stringResource(R.string.settings_holidays_calendar_no_birthdays),
+                uiLocale = uiLocale,
+                onRequestPermission = requestPermissionForListing,
+                onEnableCalendar = scrollToTop,
+            )
         }
     }
 }
@@ -671,9 +812,7 @@ private fun Collection<String>.sortedForDisplay(
 /**
  * Top-section toggle row: a checkbox + label, with an optional trailing
  * deep-link button right-aligned on the same row. Used for the Region /
- * Location / Global rows in the sources card. Calendar-sourced rows use
- * [CalendarSourceRow] instead since they need permission-prompt +
- * revocation-warning behaviour.
+ * Location / Global / Funny rows in the celebration sources card.
  *
  * The label takes `weight(1f)` so on narrow displays / long country
  * names it ellipsises before the trailing button gets squeezed off
@@ -716,9 +855,8 @@ private fun SourceRow(
  * `TextButton` whose visible text is intentionally short (just "Settings")
  * but whose accessible name is overridden via [Modifier.semantics] when a
  * [contentDescription] is supplied. Lets TalkBack / Voice Access users
- * distinguish the four otherwise-identical Settings buttons in the
- * Celebrations sources card (Region / Location / Calendar Holidays /
- * Calendar Birthdays).
+ * distinguish the otherwise-identical Settings buttons in the celebration
+ * sources card (Region / Location).
  */
 @Composable
 private fun SettingsLinkButton(
@@ -738,107 +876,27 @@ private fun SettingsLinkButton(
     ) { Text(label) }
 }
 
-/**
- * Top-section toggle row for one of the calendar-sourced theming options
- * (Calendar Holidays, Calendar Birthdays). Flips the pref through the
- * supplied callback when permission is already granted; otherwise fires
- * the `READ_CALENDAR` prompt and only flips on grant. Disabling leaves
- * permissions unchanged.
- *
- * Visual `checked` follows the stored pref directly (NOT `pref && permission`)
- * so a user who revoked READ_CALENDAR can still tap the visually-checked
- * box to disable the feature without re-granting. An inline warning row
- * surfaces the "pref says on but permission missing" state and offers a
- * re-grant.
- *
- * Revocation propagation: an ON_RESUME observer notices grant/revoke
- * transitions from system Settings and fires [onPermissionRechecked] so
- * `TodayViewModel`'s cached events list refreshes promptly instead of
- * lingering until midnight.
- */
+/** A sub-feature row: label + supporting description on the left, switch right. */
 @Composable
-private fun CalendarSourceRow(
+private fun CalendarFeatureRow(
     label: String,
+    description: String,
     checked: Boolean,
-    onSetChecked: (Boolean) -> Unit,
-    onPermissionRechecked: () -> Unit,
-    linkLabel: String,
-    linkContentDescription: String,
-    onLinkClick: () -> Unit,
+    onToggle: (Boolean) -> Unit,
 ) {
-    val context = LocalContext.current
-    var permissionGranted by remember { mutableStateOf(CalendarPermission.isGranted(context)) }
-    val currentChecked by rememberUpdatedState(checked)
-    val currentOnPermissionRechecked by rememberUpdatedState(onPermissionRechecked)
-
-    val lifecycleOwner = LocalLifecycleOwner.current
-    DisposableEffect(lifecycleOwner) {
-        val observer = LifecycleEventObserver { _, event ->
-            if (event == Lifecycle.Event.ON_RESUME) {
-                val wasGranted = permissionGranted
-                val nowGranted = CalendarPermission.isGranted(context)
-                permissionGranted = nowGranted
-                if (wasGranted != nowGranted && currentChecked) {
-                    currentOnPermissionRechecked()
-                }
-            }
-        }
-        lifecycleOwner.lifecycle.addObserver(observer)
-        onDispose { lifecycleOwner.lifecycle.removeObserver(observer) }
-    }
-
-    val launcher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.RequestPermission(),
-    ) { granted ->
-        permissionGranted = granted
-        if (granted) onSetChecked(true)
-    }
-
-    Column {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Checkbox(
-                checked = checked,
-                onCheckedChange = { wantsOn ->
-                    if (!wantsOn) {
-                        onSetChecked(false)
-                    } else if (permissionGranted) {
-                        onSetChecked(true)
-                    } else {
-                        launcher.launch(CalendarPermission.MANIFEST_PERMISSION)
-                    }
-                },
-            )
+    Row(
+        modifier = Modifier.fillMaxWidth(),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Column(modifier = Modifier.weight(1f)) {
+            Text(text = label, style = MaterialTheme.typography.bodyMedium)
             Text(
-                text = label,
-                style = MaterialTheme.typography.bodyMedium,
-                modifier = Modifier
-                    .padding(start = 8.dp)
-                    .weight(1f),
-                maxLines = 1,
-                overflow = androidx.compose.ui.text.style.TextOverflow.Ellipsis,
-            )
-            SettingsLinkButton(
-                label = linkLabel,
-                contentDescription = linkContentDescription,
-                onClick = onLinkClick,
+                text = description,
+                style = MaterialTheme.typography.bodySmall,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
             )
         }
-        if (checked && !permissionGranted) {
-            TextButton(
-                onClick = { launcher.launch(CalendarPermission.MANIFEST_PERMISSION) },
-                contentPadding = PaddingValues(horizontal = 0.dp, vertical = 0.dp),
-                modifier = Modifier.padding(start = 56.dp),
-            ) {
-                Text(
-                    text = stringResource(R.string.settings_holidays_permission_revoked),
-                    color = MaterialTheme.colorScheme.error,
-                    style = MaterialTheme.typography.bodySmall,
-                )
-            }
-        }
+        Switch(checked = checked, onCheckedChange = onToggle)
     }
 }
 
@@ -849,7 +907,7 @@ private fun CalendarSourceRow(
  * collapsed by default. Gated on the master [calendarEnabled] switch and
  * READ_CALENDAR, in that order:
  *
- *  - master off → a tappable prompt that deep-links to Calendar settings via
+ *  - master off → a tappable prompt that scrolls up to the master toggle via
  *    [onEnableCalendar];
  *  - master on, permission missing → a "grant permission" prompt that fires the
  *    system dialog via [onRequestPermission];
