+++ app/src/main/res/values/strings.xml
@@ -364,7 +364,6 @@
     <string name="settings_root_display">Display</string>
     <string name="settings_root_location">Location</string>
     <string name="settings_root_calendar">Calendar</string>
-    <string name="settings_root_holidays">Celebrations</string>
     <string name="settings_root_forecasters">Forecasters</string>
     <string name="settings_root_smart_home">Smart Home</string>
     <string name="settings_root_privacy">Privacy</string>
@@ -411,8 +410,7 @@
     <string name="settings_garment_color_dialog_close">Close</string>
 
     <string name="settings_root_location_subtitle">Automatic and Manual Location</string>
-    <string name="settings_root_calendar_subtitle">Today\'s events</string>
-    <string name="settings_root_holidays_subtitle">Holidays and Birthdays</string>
+    <string name="settings_root_calendar_subtitle">Events, Holidays, and Birthdays</string>
     <string name="settings_root_forecasters_subtitle">Weather Sources and Models</string>
     <string name="settings_root_smart_home_subtitle">Smart Displays and Home Automation</string>
     <string name="settings_root_privacy_subtitle">Crash + usage reports</string>
@@ -884,16 +882,14 @@
     <string name="settings_notification_permission_description">Without notification permission, your daily ClothesCast has nowhere to go. Grant it so ClothesCast can post tomorrow morning.</string>
     <string name="settings_notification_permission_grant">Grant notification permission</string>
 
-    <string name="settings_calendar_title">Calendar permission</string>
     <!-- Master calendar-access switch + its sub-features. The master gates all
          calendar reading; the three toggles below it pick which features use the
          calendar. Turning any sub-feature on prompts for permission and switches
          the master on automatically. Nothing read from the calendar leaves the
          device. -->
     <string name="settings_calendar_master">Use my calendar</string>
-    <string name="settings_calendar_master_description_off">Let ClothesCast read your calendar to personalise forecasts and outfits. Only event titles, dates, and times are used, and nothing leaves the device.</string>
-    <string name="settings_calendar_master_description_on">ClothesCast can read your calendar. Pick what it\'s used for below. Only event titles, dates, and times are used; nothing leaves the device.</string>
-    <string name="settings_calendar_features_title">What to use it for</string>
+    <string name="settings_calendar_master_description">Let ClothesCast read your calendars to personalise forecasts and outfits. Nothing leaves the device.</string>
+    <string name="settings_calendar_features_title">Personal calendars</string>
     <string name="settings_calendar_evening_tie_ins">Evening tie-ins</string>
     <string name="settings_calendar_evening_tie_ins_description">If you\'re going out in the evening, the morning forecast tells you if you need to bring anything extra.</string>
     <string name="settings_calendar_birthdays">Birthdays</string>
@@ -1199,12 +1195,18 @@
          "Calendar" toggles each carry a deep-link to Calendar settings;
          tapping the checkbox when permission is missing prompts for
          READ_CALENDAR. -->
+    <string name="settings_holidays_sources_title">Built-in calendars</string>
+    <!-- Expand row at the bottom of the Built-in holidays card. Reveals
+         the per-country catalogue (Global / Funny / per-country / All)
+         for users who want to flip a specific override; collapsed by
+         default so the page top stays scannable. -->
+    <string name="settings_holidays_browse_by_country">Browse by country</string>
     <string name="settings_holidays_source_region">Region (%1$s)</string>
     <string name="settings_holidays_source_location">Location (%1$s)</string>
     <string name="settings_holidays_source_global">Global</string>
     <string name="settings_holidays_source_funny">Funny</string>
-    <string name="settings_holidays_source_calendar_holidays">Calendar Holidays</string>
-    <string name="settings_holidays_source_calendar_birthdays">Calendar Birthdays</string>
+    <string name="settings_holidays_source_calendar_holidays">Calendar holidays</string>
+    <string name="settings_holidays_source_calendar_birthdays">Calendar birthdays</string>
 
     <!-- Legacy "All countries" toggle. The previous UI exposed an "All"
          checkbox that short-circuited every country on; this is dropped
@@ -1230,7 +1232,7 @@
          a grant prompt when permission is missing, a brief checking line
          while the first read runs, or a "none found" line when the synced
          calendars hold no upcoming events of that kind. -->
-    <string name="settings_holidays_calendar_enable_master">Turn on \"Use my calendar\" in Calendar settings to list upcoming celebrations.</string>
+    <string name="settings_holidays_calendar_enable_master">Turn on \"Use my calendar\" above to list upcoming celebrations.</string>
     <string name="settings_holidays_calendar_grant_permission">Grant calendar permission to list upcoming celebrations from your synced calendars.</string>
     <string name="settings_holidays_calendar_loading">Checking your calendars…</string>
     <string name="settings_holidays_calendar_no_holidays">No upcoming holidays found in your synced calendars.</string>
